Release 10.55.01
git-svn-id: http://svn.code.sf.net/p/netpbm/code/advanced@1551 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ generator/ppmcie.c
@@ -463,7 +463,7 @@ gamma_correct_rgb(const struct colorSystem * const cs,
 
 
 
-#define Sz(x) (((x) * MIN(pixcols, pixrows)) / 512)
+#define Sz(x) (((x) * (int)MIN(pixcols, pixrows)) / 512)
 #define B(x, y) ((x) + xBias), (y)
 #define Bixels(y, x) pixels[y][x + xBias]
